+++ NEWS.rst
@@ -2,6 +2,26 @@
 Version history
 ===============
 
+Plyvel 0.4
+==========
+
+Release date: 2013-06-17
+
+* Add optional 'default' argument for all ``.get()`` methods
+  (`issue #11 <https://github.com/wbolster/plyvel/issues/11>`_)
+
+
+Plyvel 0.3
+==========
+
+Release date: 2013-06-03
+
+* Fix iterator behaviour for reverse iterators using a prefix
+  (`issue #9 <https://github.com/wbolster/plyvel/issues/9>`_)
+
+* Documentation improvements
+
+
 Plyvel 0.2
 ==========